refactor: move verification methods to top level exports
